MythicBlockerFeature: Move methods to correct Util files
