+++ common/src/main/java/com/wynntils/features/user/MythicBlockerFeature.java
@@ -9,9 +9,9 @@ import com.wynntils.core.features.properties.EventListener;
 import com.wynntils.core.features.properties.FeatureInfo;
 import com.wynntils.core.features.properties.FeatureInfo.Stability;
 import com.wynntils.mc.event.ContainerCloseEvent;
-import com.wynntils.mc.utils.ItemUtils;
 import com.wynntils.mc.utils.McUtils;
 import com.wynntils.wc.utils.ContainerUtils;
+import com.wynntils.wc.utils.WynnItemMatchers;
 import com.wynntils.wc.utils.WynnUtils;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.NonNullList;
@@ -31,8 +31,8 @@ public class MythicBlockerFeature extends UserFeature {
         NonNullList<ItemStack> items = ContainerUtils.getItems(McUtils.mc().screen);
         for (int i = 0; i < 27; i++) {
             ItemStack stack = items.get(i);
-            if (!ItemUtils.isWynnItem(stack) && !ItemUtils.isUnidentifiedItem(stack)) continue;
-            if (ItemUtils.isMythic(stack)) {
+            if (!WynnItemMatchers.isWynnItem(stack) && !WynnItemMatchers.isUnidentified(stack)) continue;
+            if (WynnItemMatchers.isMythic(stack)) {
                 McUtils.sendMessageToClient(new TranslatableComponent("feature.wynntils.mythicBlocker.closingBlocked")
                         .withStyle(ChatFormatting.RED));
                 e.setCanceled(true);
